+++ packages/access-definitions/tests/data/Types/AuthorType.php
@@ -140,6 +140,7 @@ class AuthorType implements
                     true
                 )],
             ],
+            []
         );
     }