+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+* feature: add support for general update behaviors
 * feature: add support for general creation behaviors
 * feature: allow creation-configuration of ID config builders
 * feature: add support for redundant attributes/annotations